Debug dalme_app migration 0002 and everything runs cleanly

+++ dalme_app/migrations/0002_fixes_before_refactoring.py
@@ -1,12 +1,11 @@
 from django.conf import settings
 from django.db import migrations
-from django.db.utils import IntegrityError
 
 
 def fix_agent_duplicates(apps, schema_editor):
     """Update content types."""
     Agent = apps.get_model('dalme_app', 'Agent')  # noqa: N806
-    SourceCredits = apps.get_model('dalme_app', 'Source_credit')  # noqa: N806
+    SourceCredit = apps.get_model('dalme_app', 'Source_credit')  # noqa: N806
 
     control = []
     concordance = {}
@@ -24,15 +23,13 @@ def fix_agent_duplicates(apps, schema_editor):
         concordance[old_id] = Agent.objects.filter(user=user_id).order_by('creation_timestamp').first().id
     print(' \033[94m\033[1mOK\033[0m')  # noqa: T201
 
-    print('\033[95mFixing references in SourceCredits...\033[0m', end='')  # noqa: T201
-    for record in SourceCredits.objects.all():
-        if record.agent.id in concordance:
+    print('\033[95mFixing references in SourceCredit...\033[0m', end='')  # noqa: T201
+    for record in SourceCredit.objects.all():
+        if record.agent.id in concordance and not SourceCredit.objects.filter(agent__pk=record.agent.id).exists():
             new_agent = Agent.objects.get(pk=concordance[record.agent.id])
-            try:
-                record.agent = new_agent
-                record.save()
-            except IntegrityError:
-                record.delete()
+            record.agent = new_agent
+            record.save()
+            record.delete()
     print(' \033[94m\033[1mOK\033[0m')  # noqa: T201
 
     print('Deleting duplicates...', end='')  # noqa: T201
@@ -55,24 +52,24 @@ class Migration(migrations.Migration):
     operations = [
         migrations.RunSQL('ALTER TYPE varchar OWNER TO dalme;'),
         migrations.RunPython(fix_agent_duplicates),
-        migrations.RunSQL('DROP INDEX idx_16429_dalme_app_attribute_object_id_attribute_type_c83849ec;'),
+        migrations.RunSQL('DROP INDEX idx_21564_dalme_app_attribute_object_id_attribute_type_c83849ec;'),
         migrations.RunSQL(
-            'ALTER TABLE public.dalme_app_attribute ADD CONSTRAINT idx_16429_dalme_app_attribute_object_id_attribute_type_c83849ec UNIQUE(object_id, attribute_type, "value_str");',
+            'ALTER TABLE public.dalme_app_attribute ADD CONSTRAINT idx_21564_dalme_app_attribute_object_id_attribute_type_c83849ec UNIQUE(object_id, attribute_type, "value_str");',
         ),
-        migrations.RunSQL('DROP INDEX idx_16553_dalme_app_set_x_content_content_type_id_object_i_8838;'),
+        migrations.RunSQL('DROP INDEX idx_21688_dalme_app_set_x_content_content_type_id_object_i_8838;'),
         migrations.RunSQL(
-            'ALTER TABLE public.dalme_app_set_x_content ADD CONSTRAINT idx_16553_dalme_app_set_x_content_content_type_id_object_i_8838 UNIQUE(content_type_id, object_id, set_id_id);',
+            'ALTER TABLE public.dalme_app_set_x_content ADD CONSTRAINT idx_21688_dalme_app_set_x_content_content_type_id_object_i_8838 UNIQUE(content_type_id, object_id, set_id_id);',
         ),
-        migrations.RunSQL('DROP INDEX idx_16557_dalme_app_source_type_name_37ac523f_uniq;'),
+        migrations.RunSQL('DROP INDEX idx_21692_dalme_app_source_type_name_37ac523f_uniq;'),
         migrations.RunSQL(
-            'ALTER TABLE public.dalme_app_source ADD CONSTRAINT idx_16557_dalme_app_source_type_name_37ac523f_uniq UNIQUE(type, name);',
+            'ALTER TABLE public.dalme_app_source ADD CONSTRAINT idx_21692_dalme_app_source_type_name_37ac523f_uniq UNIQUE(type, name);',
         ),
-        migrations.RunSQL('DROP INDEX idx_16560_dalme_app_source_credit_source_id_agent_id_type_2bd09;'),
+        migrations.RunSQL('DROP INDEX idx_21695_dalme_app_source_credit_source_id_agent_id_type_2bd09;'),
         migrations.RunSQL(
-            'ALTER TABLE public.dalme_app_source_credit ADD CONSTRAINT idx_16560_dalme_app_source_credit_source_id_agent_id_type_2bd09 UNIQUE(source_id, agent_id, type);',
+            'ALTER TABLE public.dalme_app_source_credit ADD CONSTRAINT idx_21695_dalme_app_source_credit_source_id_agent_id_type_2bd09 UNIQUE(source_id, agent_id, type);',
         ),
-        migrations.RunSQL('DROP INDEX idx_16579_dalme_app_tasklist_group_id_slug_82602101_uniq;'),
+        migrations.RunSQL('DROP INDEX idx_21714_dalme_app_tasklist_group_id_slug_82602101_uniq;'),
         migrations.RunSQL(
-            'ALTER TABLE public.dalme_app_tasklist ADD CONSTRAINT idx_16579_dalme_app_tasklist_group_id_slug_82602101_uniq UNIQUE(group_id, slug);',
+            'ALTER TABLE public.dalme_app_tasklist ADD CONSTRAINT idx_21714_dalme_app_tasklist_group_id_slug_82602101_uniq UNIQUE(group_id, slug);',
         ),
     ]
